requirements.txt
unittest
pandas
pyyaml
python-dotenv
requests
boto3